Create two seperate modules: tasks and projects
first built module tasks in which class task is
defined. gave the required properties and exported
it. In second module projects is created. to
connect a group of tasks to a specific project,
an empty task array property is added to the
project class. Then using factory function,
createProject and getAllProject methods is build.
Now from index.js an object can be create from
projectManeger function. a project can be added
using createProject method which return an object
made by the Project class. and finally addTask
method can be used to add a new task with the
help of Task class.
[Ticket: 1]