Add getTaskTitle method to Project class
used for loop to loop through the array of objects
and get just the title of the objects and put them
in a new array and return the array. it is now
easier to display the tasks in the dom with
displayTask function and call the function every
time the add task button is clicked.

[Ticket: 3]