Add priority and bg color based on the urgency.
used select element to provide options in the
dropdown arrow to select from Low to High priority
then based on the priority the background color
changes. I used Switch and cases and attached a
new class based on the priority value and
change the color from css part for each class.

[Ticket: 7]